playlist_analysis/pyproject.toml
[project]
name = "playlist-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "seaborn>=0.13.2",
    "spotipy>=2.25.2",
    "python-dotenv>=1.0.1",
]
